sys: remove StatAtime(), StatCtime(), StatMtime() and StatATimeAsTime() utils
This removes the StatAtime(), StatCtime(), StatMtime() and StatATimeAsTime()
utilities from the sys package. These utilities are not currently used in the
containerd codebase, and a copy of these utilities can be found in
github.com/containerd/continuity/fs;

- https://github.com/containerd/continuity/blob/fa7d1622372f1a787bfe51d59183b598813ce9d9/fs/stat_darwinbsd.go
- https://github.com/containerd/continuity/blob/fa7d1622372f1a787bfe51d59183b598813ce9d9/fs/stat_linuxopenbsd.go

Given that the containerd/sys package is not designed for external consumption,
I'm not adding "deprecated" comments (or aliases to their new location).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>